+++ packages/orq-rc/src/funcs/knowledgeRetrieveDatasource.ts
@@ -109,7 +109,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "RetrieveDatasource",
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,